refactor: Refatoração das chamadas do componente de produto do client (axios) para servidor (fetch api)
