+++ frontend/services/products/productEstoqueService.ts
@@ -1,29 +1,47 @@
-import { api } from "@/app/api/api"
-import type {Category, ProductEstoque } from "@/types/product"
+import { api } from '@/app/api/api';
+import { apiFetch } from '@/app/api/server-api';
+import type {
+  Category,
+  ProductEstoque,
+  ResponseAtributos,
+} from '@/types/product';
 
 export interface IProductService {
-  getProducts(): Promise<ProductEstoque[]>
-  getCategories(): Promise<Category[]>
-  deleteProduct(id: number | string): Promise<void>
-  deleteProducts(ids: (number | string)[]): Promise<void>
+  getProducts(): Promise<ProductEstoque[]>;
+  getCategories(): Promise<Category[]>;
+  deleteProduct(id: number | string): Promise<void>;
+  deleteProducts(ids: (number | string)[]): Promise<void>;
 }
 
 export class ProductServiceEstoque implements IProductService {
   async getProducts(): Promise<ProductEstoque[]> {
-    const response = await api.get("/produtos")
-    return response.data
+    const response = await apiFetch('/produtos');
+    return response;
   }
 
   async getCategories(): Promise<Category[]> {
-    const response = await api.get("/categorias")
-    return response.data
+    const response = await apiFetch('/categorias');
+    return response;
   }
 
   async deleteProduct(id: number | string): Promise<void> {
-    await api.delete(`/produtos/${id}`)
+    await api.delete(`/produtos/${id}`);
   }
 
   async deleteProducts(ids: (number | string)[]): Promise<void> {
-    await Promise.all(ids.map((id) => this.deleteProduct(id)))
+    await Promise.all(ids.map((id) => this.deleteProduct(id)));
+  }
+
+  /**
+   * Busca os atributos por tipo de negocio
+   */
+  async getAtributosVarianteByBusiness(): Promise<ResponseAtributos[]> {
+    try {
+      const response = await apiFetch('/atributos/por-tipo-de-negocio');
+      return response.atributos;
+    } catch (error) {
+      console.error('Erro ao buscar atributos de variantes:', error);
+      return [];
+    }
   }
 }
